Добавил gateway, order, payment и delivery

+++ commerce/interaction-api/src/main/java/ru/yandex/practicum/request/ChangeProductQuantityRequest.java
@@ -15,5 +15,5 @@ public class ChangeProductQuantityRequest {
     UUID productId;
 
     @NotNull
-    Integer newQuantity;
+    Long newQuantity;
 }
